requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
dropbox==11.35.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
ply==3.11
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
stone==3.3.1
urllib3==1.26.12
Werkzeug==2.2.2
